+++ aiida_sssp_workflow/workflows/convergence/_base.py
@@ -60,6 +60,8 @@ class _BaseConvergenceWorkChain(SelfCleanWorkChain):
     def define(cls, spec):
         super().define(spec)
         # yapf: disable
+        spec.input('code', valid_type=orm.Code,
+                    help='The `pw.x` code use for the `PwCalculation`.')
         spec.input('pseudo', valid_type=UpfData, required=True,
                     help='Pseudopotential to be verified')
         spec.input('protocol', valid_type=orm.Str, required=True,
