+++ aiida_sssp_workflow/workflows/convergence/cohesive_energy.py
@@ -46,14 +46,6 @@ class ConvergenceCohesiveEnergyWorkChain(_BaseConvergenceWorkChain):
     _EVALUATE_WORKCHAIN = CohesiveEnergyWorkChain
     _MEASURE_OUT_PROPERTY = "absolute_diff"
 
-    @classmethod
-    def define(cls, spec):
-        super().define(spec)
-        # yapf: disable
-        spec.input('code', valid_type=orm.Code,
-                    help='The `pw.x` code use for the `PwCalculation`.')
-        # yapf: enable
-
     def init_setup(self):
         super().init_setup()
         self.ctx.extra_pw_parameters = {}
